requirements.txt
# EaaS Core Dependencies
# Production-grade, minimal attack surface

# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data Validation
pydantic==2.9.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Async & Networking
httpx==0.27.0
aiohttp==3.10.0
aioredis==2.0.1

# Docker Integration
docker==7.1.0

# Configuration & Utils
pyyaml==6.0.2
python-dotenv==1.0.0
tenacity==9.0.0  # Retry logic

# Security
cryptography==43.0.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.21.0

# Development & Testing (move to dev-requirements.txt in production)
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.0
mypy==1.11.0
